requirements.txt
Flask==2.3.2
gunicorn==21.2.0
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
tensorflow-cpu==2.9
textblob==0.17.1
